pyproject.toml
[project]
name = "telethon-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "beanie>=2.0.1",
    "google-auth>=2.48.0",
    "google-cloud-storage>=3.9.0",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "telethon>=1.42.0",
    "uvloop>=0.22.1",
]
